fix(types): change speaker and related fields from float to int
Changed speaker, channel, and num_words fields from float to int in
Listen V1 response types to match actual API response format. This
fixes subtitle generation issues where speakers were displayed as
"speaker 0.0" instead of "speaker 0".

Fixes #641

+++ src/deepgram/types/listen_v1response_results_utterances_item_words_item.py
@@ -11,7 +11,7 @@ class ListenV1ResponseResultsUtterancesItemWordsItem(UniversalBaseModel):
     start: typing.Optional[float] = None
     end: typing.Optional[float] = None
     confidence: typing.Optional[float] = None
-    speaker: typing.Optional[float] = None
+    speaker: typing.Optional[int] = None
     speaker_confidence: typing.Optional[float] = None
     punctuated_word: typing.Optional[str] = None
